Fix Eclipse warnings in package
git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@1044175 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/catalina/ha/backend/CollectedInfo.java
@@ -101,6 +101,6 @@ public class CollectedInfo {
         Integer ibusy  = (Integer) mBeanServer.getAttribute(objName, "currentThreadsBusy");
 
         busy = ibusy.intValue();
-        ready = imax.intValue() - ibusy;
+        ready = imax.intValue() - ibusy.intValue();
     }
 }
